chore: add use wallet provider info
